+++ changelog.md
@@ -1,3 +1,48 @@
+# 1.0.15
+
+## Tab
+- [新增] 新增`tabStyle`属性用于控制tab部分的样式
+- [修复] RN下 滚动模式下不在左边的问题
+
+## Form
+- [新增] 新增`containerProps`属性用于控制`Form.Item`容器的样式
+
+## Empty
+- [优化] 替换默认图片
+
+## Card
+- [新增] `Card.Title`组件新增样式和类属性
+
+## Calendar
+- [新增] 新增`navStyle`和`headStyle`属性用于用户控制导航部分和周导航部分样式
+- [优化] 删除默认的白色背景
+
+## Button
+- [修复] plan模式下type为custom不生效的问题
+
+## Badeg
+- [优化] 优化为点状时的默认颜色
+
+## Image
+- [新增] 新增`square`属性，控制是否显示为正方形，使用此属性需要指定width
+
+## Grade
+- [新增] 新增`type`属性，控制颜色
+
+## Radio
+- [新增] 新增`checked`属性，控制是否选中
+- [新增] 新增`onClick`属性，监听点击事件
+
+## HtmlView
+- [修复] 修复图片无法预览
+
+## Tag
+- [新增] 添加了默认样式
+
+## Text
+- [优化] 优化了Text的渲染性能
+
+
 # 1.0.8
 
 ## 更新说明